Dupla Actualizacion

+++ src/main/java/genericos/AlquilerHoy.java
@@ -3,11 +3,11 @@ import java.lang.*;
 
 public class AlquilerHoy {
 
-    private long horaRecogida;
-    private long horaEntrega;
+    private int horaRecogida;
+    private int horaEntrega;
     private int kilometrosIniciales;
 
-    public AlquilerHoy(long horaRecogida, long horaEntrega, int kilometrosIniciales) {
+    public AlquilerHoy(int horaRecogida, int horaEntrega, int kilometrosIniciales) {
         this.horaRecogida = horaRecogida;
         this.horaEntrega = horaEntrega;
         this.kilometrosIniciales = kilometrosIniciales;
